refactor(integration-tests): split TestOpenShiftScan and its non-ocp "template"
This separation will allow creating more integration tests without code
duplication in the near future (aka ansible).
